requirements.txt
## Core Dependencies
crewai==0.130.0
crewai-tools==0.47.1
fastapi==0.115.6
python-dotenv==1.0.0
uvicorn==0.32.1
python-multipart==0.0.20
PyPDF2==3.0.1

## Bonus Features - Queue Worker Model (Redis + Celery)
redis==5.2.0
celery==5.4.0
flower==2.0.1

## Bonus Features - Database Integration (PostgreSQL + SQLAlchemy)
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

## Additional Utilities
pydantic==2.10.3
httpx==0.28.1